AI_Content_Repurposing_Agent_FINAL/requirements.txt
google-genai
requests
beautifulsoup4
youtube-transcript-api
pdfplumber
python-dotenv